Move Annotation and pyramid_request into conftest.py
So that any tests in tests/h/search can use them.
